# server/src/product/base/ProductCreateInput.ts
/*
//------------------------------------------------------------------------------ 
// This code was generated by Amplication. 
// 
// Changes to this file will be lost if the code is regenerated. 
//
// There are other ways to to customize your code, see this doc to learn more
// https://docs.amplication.com/docs/how-to/custom-code
//
//------------------------------------------------------------------------------
  */
import { InputType, Field } from "@nestjs/graphql";
import { ApiProperty } from "@nestjs/swagger";
import {
  IsString,
  IsOptional,
  IsNumber,
  ValidateNested,
} from "class-validator";
import { CrategoryWhereUniqueInput } from "../../crategory/base/CrategoryWhereUniqueInput";
import { Type } from "class-transformer";
@InputType()
class ProductCreateInput {
  @ApiProperty({
    required: false,
    type: String,
  })
  @IsString()
  @IsOptional()
  @Field(() => String, {
    nullable: true,
  })
  name?: string | null;

  @ApiProperty({
    required: false,
    type: Number,
  })
  @IsNumber()
  @IsOptional()
  @Field(() => Number, {
    nullable: true,
  })
  price?: number | null;

  @ApiProperty({
    required: false,
    type: () => CrategoryWhereUniqueInput,
  })
  @ValidateNested()
  @Type(() => CrategoryWhereUniqueInput)
  @IsOptional()
  @Field(() => CrategoryWhereUniqueInput, {
    nullable: true,
  })
  category?: CrategoryWhereUniqueInput | null;
}
export { ProductCreateInput };