# swagger.json
{"openapi":"3.0.0","paths":{"/api/sign_up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentials"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/send_email_reset_password":{"post":{"operationId":"AuthController_sendEmailResetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailResetPasswordCredential"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/reset_password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordCredential"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/products":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreateInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}}},"tags":["products"],"security":[{"bearer":[]}]},"get":{"operationId":"ProductController_findMany","parameters":[{"name":"where","required":false,"in":"query","schema":{"$ref":"#/components/schemas/ProductWhereInput"}},{"name":"orderBy","required":false,"in":"query","schema":{"$ref":"#/components/schemas/ProductOrderByInput"}},{"required":false,"name":"skip","in":"query","schema":{"type":"number"}},{"required":false,"name":"take","in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getListProductDto"}}}},"403":{"description":""}},"tags":["products"],"security":[{"bearer":[]}]}},"/api/products/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["products"],"security":[{"bearer":[]}]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdateInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["products"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["products"],"security":[{"bearer":[]}]}},"/api/crategories":{"post":{"operationId":"CrategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrategoryCreateInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Crategory"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}}},"tags":["crategories"],"security":[{"bearer":[]}]},"get":{"operationId":"CrategoryController_findMany","parameters":[{"name":"where","required":false,"in":"query","schema":{"$ref":"#/components/schemas/CrategoryWhereInput"}},{"name":"orderBy","required":false,"in":"query","schema":{"$ref":"#/components/schemas/CrategoryOrderByInput"}},{"required":false,"name":"skip","in":"query","schema":{"type":"number"}},{"required":false,"name":"take","in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getListCrategoryDto"}}}},"403":{"description":""}},"tags":["crategories"],"security":[{"bearer":[]}]}},"/api/crategories/{id}":{"get":{"operationId":"CrategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Crategory"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["crategories"],"security":[{"bearer":[]}]},"patch":{"operationId":"CrategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrategoryUpdateInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Crategory"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["crategories"],"security":[{"bearer":[]}]},"delete":{"operationId":"CrategoryController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Crategory"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenException"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"tags":["crategories"],"security":[{"bearer":[]}]}},"/api/crategories/{id}/products":{"get":{"operationId":"CrategoryController_findManyProducts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":false,"in":"query","schema":{"$ref":"#/components/schemas/StringFilter"}},{"name":"name","required":false,"in":"query","schema":{"$ref":"#/components/schemas/StringNullableFilter"}},{"name":"price","required":false,"in":"query","schema":{"$ref":"#/components/schemas/FloatNullableFilter"}},{"name":"category","required":false,"in":"query","schema":{"$ref":"#/components/schemas/CrategoryWhereUniqueInput"}}],"responses":{"200":{"description":""}},"tags":["crategories"],"security":[{"bearer":[]}]},"post":{"operationId":"CrategoryController_createProducts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":""}},"tags":["crategories"],"security":[{"bearer":[]}]},"patch":{"operationId":"CrategoryController_updateProducts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["crategories"],"security":[{"bearer":[]}]},"delete":{"operationId":"CrategoryController_deleteProducts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["crategories"],"security":[{"bearer":[]}]}},"/api/_health/live":{"get":{"operationId":"HealthController_healthLive","parameters":[],"responses":{"200":{"description":""}}}},"/api/_health/ready":{"get":{"operationId":"HealthController_healthReady","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"My app","description":"\n\n## Congratulations! Your application is ready.\n  \n    Please note that all endpoints are secured with JWT Bearer authentication.Use the authentification service of supabase to authenticate.\n    (https://supabase.com/docs/gotrue/server/about#put-user)\n    Learn more in [our docs](https://docs.amplication.com)","version":"3q661ara","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserCredentials":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["email","password","role"]},"EmailResetPasswordCredential":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordCredential":{"type":"object","properties":{"access_token":{"type":"string"},"password":{"type":"string"}},"required":["access_token","password"]},"CrategoryWhereUniqueInput":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"ProductCreateInput":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"number"},"category":{"$ref":"#/components/schemas/CrategoryWhereUniqueInput"}}},"Product":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"name":{"type":"string"},"price":{"type":"number"},"category":{"$ref":"#/components/schemas/Crategory"}},"required":["id","createdAt","updatedAt"]},"Crategory":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"name":{"type":"string"},"type":{"type":"string"},"ref":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}},"required":["id","createdAt","updatedAt"]},"ForbiddenException":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"StringFilter":{"type":"object","properties":{"equals":{"type":"string"},"in":{"type":"array","items":{"type":"string"}},"notIn":{"type":"array","items":{"type":"string"}},"lt":{"type":"string"},"lte":{"type":"string"},"gt":{"type":"string"},"gte":{"type":"string"},"contains":{"type":"string"},"startsWith":{"type":"string"},"endsWith":{"type":"string"},"mode":{"type":"string","enum":["Default","Insensitive"]},"not":{"type":"string"}}},"StringNullableFilter":{"type":"object","properties":{"equals":{"type":"string"},"in":{"type":"array","items":{"type":"string"}},"notIn":{"type":"array","items":{"type":"string"}},"lt":{"type":"string"},"lte":{"type":"string"},"gt":{"type":"string"},"gte":{"type":"string"},"contains":{"type":"string"},"startsWith":{"type":"string"},"endsWith":{"type":"string"},"mode":{"type":"string","enum":["Default","Insensitive"]},"not":{"type":"string"}}},"FloatNullableFilter":{"type":"object","properties":{"equals":{"type":"number"},"in":{"type":"array","items":{"type":"number"}},"notIn":{"type":"array","items":{"type":"number"}},"lt":{"type":"number"},"lte":{"type":"number"},"gt":{"type":"number"},"gte":{"type":"number"},"not":{"type":"number"}}},"ProductWhereInput":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/StringFilter"},"name":{"$ref":"#/components/schemas/StringNullableFilter"},"price":{"$ref":"#/components/schemas/FloatNullableFilter"},"category":{"$ref":"#/components/schemas/CrategoryWhereUniqueInput"}}},"ProductOrderByInput":{"type":"object","properties":{"id":{"type":"string","enum":["Asc","Desc"]},"createdAt":{"type":"string","enum":["Asc","Desc"]},"updatedAt":{"type":"string","enum":["Asc","Desc"]},"name":{"type":"string","enum":["Asc","Desc"]},"price":{"type":"string","enum":["Asc","Desc"]},"categoryId":{"type":"string","enum":["Asc","Desc"]}}},"getListProductDto":{"type":"object","properties":{"paginatedResult":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"totalCount":{"type":"number"}},"required":["paginatedResult","totalCount"]},"NotFoundException":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"ProductUpdateInput":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"number"},"category":{"$ref":"#/components/schemas/CrategoryWhereUniqueInput"}}},"CrategoryCreateInput":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"ref":{"type":"string"}}},"CrategoryWhereInput":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/StringFilter"},"name":{"$ref":"#/components/schemas/StringNullableFilter"},"type":{"$ref":"#/components/schemas/StringNullableFilter"},"ref":{"$ref":"#/components/schemas/StringNullableFilter"}}},"CrategoryOrderByInput":{"type":"object","properties":{"id":{"type":"string","enum":["Asc","Desc"]},"createdAt":{"type":"string","enum":["Asc","Desc"]},"updatedAt":{"type":"string","enum":["Asc","Desc"]},"name":{"type":"string","enum":["Asc","Desc"]},"type":{"type":"string","enum":["Asc","Desc"]},"ref":{"type":"string","enum":["Asc","Desc"]}}},"getListCrategoryDto":{"type":"object","properties":{"paginatedResult":{"type":"array","items":{"$ref":"#/components/schemas/Crategory"}},"totalCount":{"type":"number"}},"required":["paginatedResult","totalCount"]},"CrategoryUpdateInput":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"ref":{"type":"string"}}}}}}